rigid: a rigid body orbital simulation (megastructure) branch
First we demonstrate a skyhook (with cross shape for additional stability).

What we already learn with it is that it's stable with tidally locked
rotation (revolves at the same rate of orbiting), which is often observed
with natural celestial bodies.  It is unexpected because tidal lock is
normally explained with internal friction due to non-uniform gravitational
field, but actually it works with rigid bodies (which don't have internal
friction).